Refactor code with CR's recommandations
