+++ create_rental.rb
